Rename files to OpenAL specific. Abstract methods later
