+++ engine/audio/src/openal_device.cpp
@@ -1,0 +1,17 @@
+#include "audio/openal_device.h"
+
+namespace Ragnarok {
+
+        Device::Device();
+        
+        Device::~Device();
+
+        void Device::changePlaybackDevice(const ALChar *deviceName) {
+        
+        }
+
+        void Device::changeCaptureDevice(const ALChar *deviceName) {
+        
+        }
+
+}
